move BlockPosUtil into the general util package
