+++ src/main/java/com/gregtechceu/gtceu/utils/BlockPosUtil.java
@@ -1,4 +1,4 @@
-package com.gregtechceu.gtceu.utils.fakelevel;
+package com.gregtechceu.gtceu.utils;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.phys.Vec3;
